feat: [M3-7036] Naming conventions, component splitting and BarPercent
