+++ packages/manager/src/components/BarPercent/BarPercent.tsx
@@ -8,6 +8,12 @@ import { isPropValid } from 'src/utilities/isPropValid';
 export interface BarPercentProps {
   /** Additional css class to pass to the component */
   className?: string;
+  /**
+   * Applies a gradient to the filler bar to show
+   * when getting close to the limit. (green to red)
+   * Orange will show at 75% and red at 90%.
+   */
+  gradientFiller?: boolean;
   /** Applies styles to show that the value is being retrieved. */
   isFetchingValue?: boolean;
   /** The maximum allowed value and should not be equal to min. */
@@ -29,6 +35,7 @@ export interface BarPercentProps {
 export const BarPercent = (props: BarPercentProps) => {
   const {
     className,
+    gradientFiller,
     isFetchingValue,
     max,
     narrow,
@@ -37,6 +44,8 @@ export const BarPercent = (props: BarPercentProps) => {
     value,
     valueBuffer,
   } = props;
+
+  const fillerRelativePct = (100 / value) * 100;
 
   return (
     <StyledDiv className={`${className}`}>
@@ -48,6 +57,8 @@ export const BarPercent = (props: BarPercentProps) => {
             ? 'buffer'
             : 'determinate'
         }
+        fillerRelativePct={fillerRelativePct}
+        gradientFiller={gradientFiller}
         narrow={narrow}
         rounded={rounded}
         sx={sx}
@@ -72,18 +83,29 @@ const StyledDiv = styled('div')({
 const StyledLinearProgress = styled(LinearProgress, {
   label: 'StyledLinearProgress',
   shouldForwardProp: (prop) => isPropValid(['rounded'], prop),
-})<Partial<BarPercentProps>>(({ theme, ...props }) => ({
-  '& .MuiLinearProgress-bar2Buffer': {
-    backgroundColor: '#99ec79',
-  },
-  '& .MuiLinearProgress-barColorPrimary': {
-    backgroundColor: '#5ad865',
-  },
-  '& .MuiLinearProgress-dashed': {
-    display: 'none',
-  },
-  backgroundColor: theme.color.grey2,
-  borderRadius: props.rounded ? theme.shape.borderRadius : undefined,
-  padding: props.narrow ? 8 : 12,
-  width: '100%',
-}));
+})<Partial<BarPercentProps> & { fillerRelativePct: number }>(
+  ({ theme, ...props }) => ({
+    '& .MuiLinearProgress-bar2Buffer': {
+      backgroundColor: '#99ec79',
+    },
+    '& .MuiLinearProgress-barColorPrimary': {
+      backgroundColor: '#5ad865',
+    },
+    '& .MuiLinearProgress-dashed': {
+      display: 'none',
+    },
+    backgroundColor: theme.color.grey2,
+    borderRadius: props.rounded ? theme.shape.borderRadius : undefined,
+    padding: props.narrow ? 8 : 12,
+    width: '100%',
+    ...(props.gradientFiller && {
+      '& .MuiLinearProgress-bar': {
+        background: `linear-gradient(90deg, #5ad865 0%, #5ad865 75%, orange 90%, red);`,
+        transform: 'none !important',
+        width: `${props.fillerRelativePct}%`,
+      },
+      backgroundColor: 'rgb(231, 231, 231)',
+      width: `${props.value}%`,
+    }),
+  })
+);
