+++ packages/manager/src/components/TransferDisplay/TransferDisplayUsage.tsx
@@ -1,9 +1,9 @@
 import Grid from '@mui/material/Unstable_Grid2';
 import * as React from 'react';
 
+import BarPercent from 'src/components/BarPercent';
 import { Typography } from 'src/components/Typography';
 
-import { StyledBarPercent } from './TransferDisplay.styles';
 import { formatPoolUsagePct } from './utils';
 
 export interface TransferDisplayUsageProps {
@@ -17,7 +17,6 @@ export const TransferDisplayUsage = React.memo(
     const { pullUsagePct, quota, used } = props;
     // Don't display usage, quota, or bar percent if the network transfer pool is empty (e.g. account has no resources).
     const isEmptyPool = quota === 0;
-    const fillerRelativePct = (100 / Math.ceil(pullUsagePct)) * 100;
 
     return (
       <>
@@ -54,10 +53,9 @@ export const TransferDisplayUsage = React.memo(
           </Grid>
         </Grid>
         {!isEmptyPool && (
-          <StyledBarPercent
-            fillerRelativePct={fillerRelativePct}
+          <BarPercent
+            gradientFiller
             max={100}
-            pullUsagePct={pullUsagePct}
             rounded
             value={Math.ceil(pullUsagePct)}
           />
