+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeNetworking/NetworkingSummaryPanel/NetworkTransfer.tsx
@@ -1,11 +1,7 @@
 import { getLinodeTransfer } from '@linode/api-v4/lib/linodes';
-import Grid from '@mui/material/Unstable_Grid2';
-import { styled, useTheme } from '@mui/material/styles';
+import { useTheme } from '@mui/material/styles';
 import * as React from 'react';
 
-import BarPercent from 'src/components/BarPercent';
-import { CircleProgress } from 'src/components/CircleProgress';
-import { Notice } from 'src/components/Notice/Notice';
 import { TextTooltip } from 'src/components/TextTooltip';
 import { Typography } from 'src/components/Typography';
 import { useAPIRequest } from 'src/hooks/useAPIRequest';
@@ -16,6 +12,8 @@ import { useTypeQuery } from 'src/queries/types';
 import { MONTHLY_NETWORK_TRANSFER_TOOLTIP_MESSAGE } from 'src/utilities/pricing/constants';
 import { isLinodeInDynamicPricingDC } from 'src/utilities/pricing/linodes';
 import { readableBytes } from 'src/utilities/unitConversions';
+
+import { TransferContent } from './TransferContent';
 
 import type { Region } from '@linode/api-v4';
 
@@ -58,7 +56,7 @@ export const NetworkTransfer = React.memo((props: Props) => {
 
   const dynamicDCTooltip = isDynamicPricingDC ? (
     <TextTooltip
-      displayText={currentRegion?.label || ''}
+      displayText={`${currentRegion?.label} datacenter` || ''}
       minWidth={275}
       placement="right-end"
       tooltipText={MONTHLY_NETWORK_TRANSFER_TOOLTIP_MESSAGE}
@@ -92,145 +90,3 @@ export const NetworkTransfer = React.memo((props: Props) => {
     </div>
   );
 });
-
-// =============================================================================
-// TransferContent (With loading and error states)
-// =============================================================================
-interface ContentProps {
-  accountBillableInGB: number;
-  accountQuotaInGB: number;
-  error: boolean;
-  linodeLabel: string;
-  linodeUsedInGB: number;
-  loading: boolean;
-  totalUsedInGB: number;
-}
-
-const TransferContent = (props: ContentProps) => {
-  const {
-    accountQuotaInGB,
-    error,
-    linodeLabel,
-    linodeUsedInGB,
-    loading,
-    totalUsedInGB,
-    // accountBillableInGB
-  } = props;
-  const theme = useTheme();
-
-  /**
-   * In this component we display three pieces of information:
-   *
-   * 1. Account-level transfer quota for this month
-   * 2. The amount of transfer THIS Linode has used this month
-   * 3. The remaining transfer on your account this month
-   */
-
-  const linodeUsagePercent = calculatePercentageWithCeiling(
-    linodeUsedInGB,
-    accountQuotaInGB
-  );
-
-  const totalUsagePercent = calculatePercentageWithCeiling(
-    totalUsedInGB,
-    accountQuotaInGB
-  );
-
-  const remainingInGB = Math.max(accountQuotaInGB - totalUsedInGB, 0);
-
-  if (error) {
-    return (
-      <Notice
-        text={
-          'Network transfer information for this Linode is currently unavailable.'
-        }
-        important
-        spacingBottom={0}
-        variant="error"
-      />
-    );
-  }
-
-  if (loading) {
-    return (
-      <Grid container justifyContent="center">
-        <Grid>
-          <CircleProgress mini />
-        </Grid>
-      </Grid>
-    );
-  }
-
-  return (
-    <div>
-      <BarPercent
-        max={100}
-        rounded
-        sx={{ marginBottom: `${theme.spacing(0.5)}px` }}
-        value={Math.ceil(linodeUsagePercent)}
-        valueBuffer={Math.ceil(totalUsagePercent)}
-      />
-      <StyledGreenTypography>
-        <span>
-          {linodeLabel} ({linodeUsedInGB} GB)
-        </span>
-      </StyledGreenTypography>
-      <StyledGreyTypography>
-        <span>Remaining ({remainingInGB} GB)</span>
-      </StyledGreyTypography>
-      {/* @todo: display overages  */}
-    </div>
-  );
-};
-
-const sxLegendItemBefore = {
-  borderRadius: 5,
-  content: '""',
-  height: 20,
-  marginRight: 10,
-  width: 20,
-};
-
-const sxLegendItem = {
-  alignItems: 'center',
-  display: 'flex',
-  marginTop: 10,
-};
-
-const StyledGreyTypography = styled(Typography, {
-  label: 'StyledGreyTypography',
-})(({ theme }) => ({
-  ...sxLegendItem,
-  '&  span': {
-    flex: 1,
-  },
-  '&:before': {
-    ...sxLegendItemBefore,
-    backgroundColor: theme.color.grey2,
-  },
-}));
-
-const StyledGreenTypography = styled(Typography, {
-  label: 'StyledGreenTypography ',
-})({
-  ...sxLegendItem,
-  '&  span': {
-    flex: 1,
-  },
-  '&:before': {
-    ...sxLegendItemBefore,
-    backgroundColor: '#5ad865',
-  },
-});
-
-// =============================================================================
-// Utilities
-// =============================================================================
-
-// Get the percentage of a value in relation to a given target. Caps return value at 100%.
-export const calculatePercentageWithCeiling = (
-  value: number,
-  target: number
-) => {
-  return target > value ? 100 - ((target - value) * 100) / target : 100;
-};
